moved home page to its own component
